(ok-89s52): fix build errors caued by math functions
Change list
1. 'abs' is declared in 'stdlib.h'
2. SDCC does not support double-precision math functions. replace them
   with 'single-precision's.

+++ Exp08_4/Exp08_4.c
@@ -65,7 +65,7 @@ void main(void)
       LCD_command(0xC5);
       LCD_2d1(x);                              // display argument
       LCD_command(0xCB);
-      LCD_1d3(sqrt(x));                        // display sqrt
+      LCD_1d3(sqrtf(x));                       // display sqrt
       Delay_ms(2000);
     }
   }
